+++ src/main/java/com/revature/autosurvey/users/UsersApplication.java
@@ -1,4 +1,4 @@
-package com.revature.autosurvey;
+package com.revature.autosurvey.users;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
