+++ src/main/java/com/revature/autosurvey/users/repos/UserRepo.java
@@ -1,10 +1,10 @@
-package com.revature.autosurvey.repos;
+package com.revature.autosurvey.users.repos;
 
 import java.util.UUID;
 
 import org.springframework.data.cassandra.repository.ReactiveCassandraRepository;
 
-import com.revature.autosurvey.beans.User;
+import com.revature.autosurvey.users.beans.User;
 
 public interface UserRepo extends ReactiveCassandraRepository<User, UUID>{
