+++ src/main/java/com/revature/autosurvey/users/services/UserService.java
@@ -1,5 +1,21 @@
 package com.revature.autosurvey.users.services;
 
-public class UserService {
+import com.revature.autosurvey.users.beans.User;
 
+import reactor.core.publisher.Flux;
+import reactor.core.publisher.Mono;
+
+public interface UserService {
+	
+	Flux<User> getAllUsers();
+	
+	Mono<User> getUserByEmail(String email);
+	
+	Mono<User> addUser(User user);
+	
+	Mono<User> updateUser(User user);
+	
+	Mono<User> getUserById(String ID);
+	
+	Mono<User> deleteUser(String ID);
 }
