+++ src/main/java/com/revature/autosurvey/users/services/UserServiceImp.java
@@ -1,6 +1,6 @@
-package com.revature.autosurvey.services;
+package com.revature.autosurvey.users.services;
 
-import com.revature.autosurvey.beans.User;
+import com.revature.autosurvey.users.beans.User;
 
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
