+++ src/main/java/com/revature/autosurvey/users/users/controllers/UserController.java
@@ -1,4 +1,4 @@
-package com.revature.autosurvey.users.controllers;
+package com.revature.autosurvey.users.users.controllers;
 
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -7,7 +7,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.revature.autosurvey.beans.User;
+import com.revature.autosurvey.users.beans.User;
 
 @RestController
 @RequestMapping(value = "/users")
