+++ src/main/java/com/revature/autosurvey/users/users/services/UserService.java
@@ -1,0 +1,5 @@
+package com.revature.autosurvey.users.users.services;
+
+public class UserService {
+
+}
